#245 coderabbit指摘修正
